:sparkles: Added CRUD operations for User Management
This commits adds simple index, edit and create websites for the
user(admins). The index page displays the users with pagination and also
there is the feature for searching, filtering with status and order by
name, in this admin index page there is the option to add new users and
also to edit them.

Added also the business logic to the admin controller, updated the
requests validation forms